Modified Fibonacci.cpp and Prime.cpp

+++ Fibonacci.cpp
@@ -11,9 +11,9 @@ using namespace std;
 
 int main() {
 
- int previous = 1;
- int current = 1;
- int sum;
+ long double previous = 1;
+ long double current = 1;
+ long double sum;
  int count;
 
  cout << "How many fibonacci numbers do you want to see?" << endl;
